backend/requirements.txt
# Backend Requirements for Sensei Face Recognition System

# Flask Framework
Flask==3.1.2
Flask-CORS==3.0.10
Flask-SQLAlchemy==3.0.2

# Database
PyMySQL==1.1.2
SQLAlchemy==2.0.46

# Face Recognition
opencv-python==4.8.1.78
dlib==19.24.0
face-recognition==1.3.0
numpy>=1.25.2,<2.0.0
Pillow>=10.3.0

# Utilities
python-dotenv==1.0.0
werkzeug==3.1.0

# Optional: Image processing
scikit-image==0.22.0
